Improved runtime API interface
Before, the emulator was duplicating code from the runtime API
implementations. We now offer this code from a `runtime_api` module from
the runtime. We also move everything related to the Substrate runtime
implementation (that is the `impl_runtime_apis!` call ) to this module
to improve encapsulation.
